test: Alternative source address for rpc_bind
Instead of requiring a second interface for the rpcallow tests,
use a second loopback address. This should work just as well
but removes the requirement for binding on non-localhost.

This needs support in the test framework at the RPC/HTTP level
for binding to a different address before outgoing connection
(connecting to a different interface IP does that part
automatically...).