guix: Create linker maps, and ship with debug artifacts
The linker map (`<BINARY>.map`) provides accounting of where every little
fragment of the binary comes from and ends up.

This is extremely useful for binary comparison, and other things
such as statistics on binary size.

Enable it on GNU linker (linux, windows) only--it doesn't look
like the Apple linker likes this flag, and we don't ship debug
information for that anyhow.